Adds compile-time validation (#2)
Adds a compiler-plugin that runs the smithy validators on smithy-models translated from `API` definitions found in the compiled code. 

This plugin is bundled in an artifact using assembly, as the build tools (SBT, scala-cli, ...) do not pull the transitive dependencies of compiler plugins. The bundled dependencies are, in particular, classgraph, smithy, smithy4s, smithy4s-deriving.